Move Turfgame URL methods to own class.
