+++ src/main/java/org/joelson/turf/turfgame/apiv4/Rounds.java
@@ -1,8 +1,9 @@
 package org.joelson.turf.turfgame.apiv4;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import org.joelson.turf.turfgame.util.TurfgameURLReader;
 import org.joelson.turf.util.JacksonUtil;
-import org.joelson.turf.util.URLReader;
+import org.joelson.turf.util.URLReader.Response;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -18,7 +19,7 @@ public final class Rounds {
     }
 
     public static List<Round> readRounds() throws IOException {
-        URLReader.Response response = URLReader.getTurfgameRequest(ROUNDS_REQUEST);
+        Response response = TurfgameURLReader.getTurfgameRequest(ROUNDS_REQUEST);
         if (response.status() != HttpURLConnection.HTTP_OK) {
             System.err.printf("Response status: %d, request: %s", response.status(), ROUNDS_REQUEST);
         }
