+++ src/main/java/org/joelson/turf/turfgame/apiv4/Users.java
@@ -4,8 +4,9 @@ import com.fasterxml.jackson.core.JsonEncoding;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import org.joelson.turf.turfgame.util.TurfgameURLReader;
 import org.joelson.turf.util.JacksonUtil;
-import org.joelson.turf.util.URLReader;
+import org.joelson.turf.util.URLReader.Response;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -47,7 +48,7 @@ public final class Users {
         generator.writeEndArray();
         generator.close();
         String requestJSON = stream.toString(StandardCharsets.UTF_8);
-        URLReader.Response response = URLReader.postTurfgameRequest(USERS_REQUEST, requestJSON);
+        Response response = TurfgameURLReader.postTurfgameRequest(USERS_REQUEST, requestJSON);
         if (response.status() != HttpURLConnection.HTTP_OK) {
             System.err.printf("Response status: %d, request URL: %s, JSON: %s%n", response.status(), USERS_REQUEST,
                     requestJSON);
