+++ src/main/java/org/joelson/turf/turfgame/apiv5/Zones.java
@@ -1,8 +1,9 @@
 package org.joelson.turf.turfgame.apiv5;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import org.joelson.turf.turfgame.util.TurfgameURLReader;
 import org.joelson.turf.util.JacksonUtil;
-import org.joelson.turf.util.URLReader;
+import org.joelson.turf.util.URLReader.Response;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -30,7 +31,7 @@ public class Zones {
     }
 
     private static String getAllZonesJSON() throws IOException {
-        URLReader.Response response = URLReader.getTurfgameRequest(ALL_ZONES_REQUEST);
+        Response response = TurfgameURLReader.getTurfgameRequest(ALL_ZONES_REQUEST);
         if (response.status() != HttpURLConnection.HTTP_OK) {
             System.err.printf("Response status: %d, request: %s", response.status(), ALL_ZONES_REQUEST);
         }
